fix(ESSNTL-726): move change of display name to default details reducer
