+++ src/store/entityDetails.js
@@ -39,6 +39,48 @@ function showTagsPending(state) {
     };
 }
 
+function updateAnsibleName(state, { meta }, useOrigValue) {
+    const value = useOrigValue ? meta?.origValue : meta?.value;
+    return {
+        ...state,
+        ...state.rows && {
+            rows: state.rows.map((row) => row.id === meta?.id ? ({
+                ...row,
+                // eslint-disable-next-line camelcase
+                ansible_host: value
+            }) : row)
+        },
+        ...state.entity && {
+            entity: {
+                ...state.entity,
+                // eslint-disable-next-line camelcase
+                ansible_host: value
+            }
+        }
+    };
+}
+
+export function updateEntity(state, { meta }, useOrigValue) {
+    const value = useOrigValue ? meta?.origValue : meta?.value;
+    return {
+        ...state,
+        ...state.rows && {
+            rows: state.rows.map((row) => row.id === meta?.id ? ({
+                ...row,
+                // eslint-disable-next-line camelcase
+                display_name: value
+            }) : row)
+        },
+        ...state.entity && {
+            entity: {
+                ...state.entity,
+                // eslint-disable-next-line camelcase
+                display_name: value
+            }
+        }
+    };
+}
+
 export default {
     [ACTION_TYPES.LOAD_ENTITIES_PENDING]: () => entityDefaultState,
     [ACTION_TYPES.LOAD_ENTITY_PENDING]: entityDetailPending,
@@ -49,5 +91,9 @@ export default {
     [ACTION_TYPES.LOAD_TAGS_FULFILLED]: showTags,
     [TOGGLE_TAG_MODAL]: toggleTagModalReducer,
     [TOGGLE_DRAWER]: toggleDrawer,
+    [ACTION_TYPES.UPDATE_DISPLAY_NAME_PENDING]: updateEntity,
+    [ACTION_TYPES.SET_ANSIBLE_HOST_PENDING]: updateAnsibleName,
+    [ACTION_TYPES.UPDATE_DISPLAY_NAME_ERROR]: (state, payload) => updateEntity(state, payload, true),
+    [ACTION_TYPES.SET_ANSIBLE_HOST_ERROR]: (state, payload) => updateAnsibleName(state, payload, true),
     ...systemIssuesReducer
 };
